+++ virus_executor_service/include/ves/ves_protocol.h
@@ -5,7 +5,7 @@
 
 namespace VirusExecutorService {
 
-enum class VesOpcode : uint16_t {
+enum class VesOpcode : uint8_t {
     ScanFile = 102,
 };
